Remove download count retention period from README
In order to view the impact of changes such as OS releases over a longer
period of time, we have removed the automatic expiry of the anonymous
download counts from the rpi-imager stats service.

For reference, the download counts contain only the following data:

    > ZREVRANGE image:2021-03-12 0 1 WITHSCORES
    1) Raspberry Pi OS (32-bit)
    2) 1234
    3) Raspberry Pi OS Lite (32-bit)
    4) 567

Update the README section about telemetry in order to reflect the
change.