Linux/udisks2/image customization: fix issue with mounting NVMe drives
When using image customization with an NVMe drive under RPI OS on a CM4,
the system will prompt non-root users for password when mounting the FAT file system.
(Passwordless mounting does is available for removable storage,
but the NVMe drive qualifies as internal storage instead).

- The mount command we send to udisks2 over dbus2 may timeout
before the user completed the password prompt.
Raise timeout to counter this.

- Print mount error details to console
- On mount error report error mounting in GUI instead of
  "/config.txt does not exist" (which happened when mountDevice()
  returned empty string)