i18n: Advanced settings -> OS customization
Per a request from the Raspberry Pi Communications team, change
'Advanced Settings' and 'Image customization' to read 'OS
customization'.

This is to avoid amibuguity in the English, where Image_r_ and Image may
be conflated when reading at speed.

Regrettably, this breaks the existing translations. To prevent
confusion, explicitly delete the existing translations - as they're almost
certainly not valid with the reworded meaning.