imagewriter: Use QCoreApplication::translate
tr() is not safe for strings that you move between contexts,
as it fails to provide a mechanism for specifying the context
of the string.

Regenerate the translation files, and use QCoreApplication to locate
the real translations.

Additionally, bodge a race condition between translator replacement
and OS list fetching, by marking the OS list as something to be
re-calculated on a translator change.

Fixes #799