Split OptionsPopup.qml into multiple files
One file per tab.

Besides reducing the size of OptionsPopup.qml it's also nice that
each tab has an "interface" with public properties while hiding
the private impl.

For OptionsMiscTab.qml I've improved semantics by only exposing
a bool telemetryEnabled instead of the whole check-box, which
was very verbose in the caller site.

In a followup the 3 tabs will inherit from a base component,
so we can share the scrolling behaviour properties