OS customisation: Remove 'Show Password'.
This feature causes more confusion that it does help.

After saving the settings, we hash the password to prevent offline plaintext recovery, and do so in a format that wpa_supplicant will be happy with.

Users, however, continue to believe this is memory corruption - and not unfairly ask why they can't see their actual password (unaware that the hash is a one-way function).

The only other viable option is to make this box appear in a dynamic manner, however dynamic UIs are hard to describe and work with - and often rely on context that is non-obvious.

So, just remove it. If you type a password, you'll still see each character in the clear as you type it.