fix(ImComboBox): Add platform-specific scroll helpers
Previously, we relied entirely on Qt to tell us if scroll events were
inverted or no - and that worked great on macOS.

On Windows, however, .inverted was _never set_, even when the Control
Panel was set to invert the scrolling direction.

So don't rely on Qt, and introduce a platform quirk to get the correct
scrolling direction for ImComboBox.